Uses webpack for node build
