+++ src/shared.js
@@ -1,4 +1,4 @@
-export const pathJoin = (...paths) => {
+export function pathJoin (...paths) {
   let newPath = `${paths.join('/')}`.replace(/\/{2,}/g, '/')
   if (newPath !== '/') {
     newPath = newPath.replace(/\/$/g, '')
